+++ .github/workflows/release.yml
@@ -1,17 +1,22 @@
 name: Release
 
 on:
+  workflow_dispatch:
   schedule:
     - cron: "0 */8 * * *"
 
 jobs:
   lint:
-    name: lint
+    name: release
     runs-on: ubuntu-latest
     steps:
-    - uses: actions/checkout@master
+    - uses: actions/checkout@v2
       with:
         token: ${{ secrets.GITHUB_PAT_FROM_ZEKE }}
+    - name: Use Node.js 14.x
+      uses: actions/setup-node@v2
+      with:
+        node-version: 14.x
     - name: npm ci
       run: npm ci
     - name: release
@@ -21,8 +26,11 @@ jobs:
         git config --global user.email "github-actions@users.noreply.github.com"
 
         npm run update
-        npm test
+        npm run test
         [[ `git status --porcelain` ]] || exit
+
+        # update stats
+        npm run stats
 
         echo "//registry.npmjs.org/:_authToken=${{ secrets.NPM_TOKEN }}" > ~/.npmrc
