+++ .gitignore
@@ -2,3 +2,5 @@
 .npmrc
 .cache/
 node_modules
+.idea/
+.DS_Store
