+++ README.md
@@ -1,0 +1,107 @@
+# all-the-package-repos
+
+All the repository URLs in the npm registry as an object whose keys are package names and values are URLs.
+
+This package weighs in at about 30MB.
+
+## Stats
+
+<!-- stats -->
+Packages | Count | Percentage
+-------- | ----- | ----------
+With repository in package.json | 1096618 | 63.74%
+On GitHub | 1075533 |  98.08%
+On BitBucket | 1183 |  0.11%
+On GitLab | 1288 |  0.12%
+<!-- /stats -->
+
+## Installation
+
+```sh
+npm install all-the-package-repos --save
+```
+
+## Usage
+
+```js
+repos = require('all-the-package-repos')
+
+repos.express
+// https://github.com/expressjs/express
+```
+
+See [example.js](example.js) for more usage details.
+
+GitHub URLs are normalized to their `https` form using
+[github-url-to-object](http://ghub.io/github-url-to-object):
+
+- `git@github.com:foo/bar.git` becomes `https://github.com/foo/bar`
+- `foo/bar` becomes `https://github.com/foo/bar`
+- [etc...](http://ghub.io/github-url-to-object)
+
+### Repository Hostnames
+
+For the curious, there's a submodule that collects all the hostnames of all the
+repository URLS:
+
+```js
+require('./hostnames').slice(0,10)
+
+[ 
+  { value: 'github.com', count: 452768 },
+  { value: 'bitbucket.org', count: 553 },
+  { value: 'git.oschina.net', count: 219 },
+  { value: 'gitlab.com', count: 116 },
+  { value: 'git.coding.net', count: 114 },
+  { value: 'archive.voodoowarez.com', count: 81 },
+  { value: 'gitee.com', count: 60 },
+  { value: 'gitlab.baidu.com', count: 49 },
+  { value: 'git-wip-us.apache.org', count: 38 },
+  { value: 'gitlab.alibaba-inc.com', count: 36 }
+]
+```
+
+It also has a CLI:
+
+```sh
+all-the-package-repo-hostnames | head -n 10
+
+github.com                                        452768
+bitbucket.org                                     553
+git.oschina.net                                   219
+gitlab.com                                        116
+git.coding.net                                    114
+archive.voodoowarez.com                           81
+gitee.com                                         60
+gitlab.baidu.com                                  49
+git-wip-us.apache.org                             38
+gitlab.alibaba-inc.com                            36
+```
+
+## Tests
+
+```sh
+npm install
+npm test
+```
+
+## Dependencies
+
+- [count-array-values](https://ghub.io/count-array-values): Count the instances of each value in an array
+- [pad](https://ghub.io/pad): Left and right string padding
+
+## Dev Dependencies
+
+- [bitbucket-url-to-object](https://ghub.io/bitbucket-url-to-object): Extract user, repo, and other interesting properties from Bitbucket URLs
+- [chai](https://ghub.io/chai): BDD/TDD assertion library for node.js and the browser. Test framework agnostic.
+- [github-url-to-object](https://ghub.io/github-url-to-object): Extract user, repo, and other interesting properties from GitHub URLs
+- [is-url](https://ghub.io/is-url): Check whether a string is a URL.
+- [mocha](https://ghub.io/mocha): simple, flexible, fun test framework
+- [nano](https://ghub.io/nano): The official CouchDB client for Node.js
+- [standard](https://ghub.io/standard): JavaScript Standard Style
+
+## License
+
+MIT
+
+_Generated by [package-json-to-readme](https://github.com/zeke/package-json-to-readme)_
