+++ test.js
@@ -17,11 +17,11 @@ describe('repos', () => {
     expect(repos.express).to.equal('https://github.com/expressjs/express')
   })
 
-  it('is always a URL', function () {
-    this.timeout(10 * 1000)
+  it('is always a URL or null', function () {
+    this.timeout(30 * 1000)
     const urls = Object.values(repos)
     urls.forEach(url => {
-      expect(isUrl(url), `${url}`).to.eq(true)
+      expect(isUrl(url) || url === null, `${url}`).to.eq(true)
     })
   })
